Edita Orlova: add new test into group class
